[Chore] Migrate to Xcode 11.3 (#155)
* Pin components

* Update project settings

* Fix deprecation warning

* Fix deprecation warning

* Fix test compile error

* Fix mutability warning

* Fix linker warning (incorrect method override)

* Pin azure branch

* Use swift 5.1

* Update project-common.xcconfig

* Update azure-pipelines.yml

* bump

+++ Tests/Source/TransportEncoding/ZMUpdateEventTests.m
@@ -98,7 +98,7 @@
 
         // then
         XCTAssertNotNil(event);
-        XCTAssertEqual([types[key] intValue], event.type);
+        XCTAssertEqual([types[key] intValue], (int) event.type);
     }
 
 }
